+++ unity-ggjj/Assets/Scripts/SaveFiles/SettingsSaveDataManager.cs
@@ -1,29 +1,28 @@
 using UnityEngine;
 using UnityEngine.Audio;
-using UnityEngine.Serialization;
 
 namespace SaveFiles
 {
     public class SettingsSaveDataManager : MonoBehaviour
     {
-        [SerializeField] private AudioMixer audioMixerGroup;
+        [SerializeField] private AudioMixer _audioGroup;
 
         private void Start()
         {
             var settings = PlayerPrefsProxy.Load<SettingsSaveData>(SettingsSaveData.Key);
 
-            audioMixerGroup.SetFloat("MasterVolume", settings.GameSettings.AudioSettings.Master);
-            audioMixerGroup.SetFloat("MusicVolume", settings.GameSettings.AudioSettings.Music);
-            audioMixerGroup.SetFloat("SFXVolume", settings.GameSettings.AudioSettings.Sfx);
+            _audioGroup.SetFloat("MasterVolume", settings.GameSettings.AudioSettings.Master);
+            _audioGroup.SetFloat("MusicVolume", settings.GameSettings.AudioSettings.Music);
+            _audioGroup.SetFloat("SFXVolume", settings.GameSettings.AudioSettings.Sfx);
         }
 
         public void SaveAudioChanges()
         {
             var settings = PlayerPrefsProxy.Load<SettingsSaveData>(SettingsSaveData.Key);
 
-            audioMixerGroup.GetFloat("MasterVolume", out settings.GameSettings.AudioSettings.Master);
-            audioMixerGroup.GetFloat("MusicVolume", out settings.GameSettings.AudioSettings.Music);
-            audioMixerGroup.GetFloat("SFXVolume", out settings.GameSettings.AudioSettings.Sfx);
+            _audioGroup.GetFloat("MasterVolume", out settings.GameSettings.AudioSettings.Master);
+            _audioGroup.GetFloat("MusicVolume", out settings.GameSettings.AudioSettings.Music);
+            _audioGroup.GetFloat("SFXVolume", out settings.GameSettings.AudioSettings.Sfx);
 
             PlayerPrefsProxy.Save(settings);
         }
